Improve code quality for find
