+++ src/main/java/seedu/clinkedin/model/AddressBook.java
@@ -111,7 +111,6 @@ public class AddressBook implements ReadOnlyAddressBook {
     @Override
     public String toString() {
         return persons.asUnmodifiableObservableList().size() + " persons";
-        // TODO: refine later
     }
 
     @Override
