+++ src/main/java/seedu/clinkedin/model/VersionedAddressBook.java
@@ -204,7 +204,6 @@ public class VersionedAddressBook extends AddressBook {
     @Override
     public String toString() {
         return persons.asUnmodifiableObservableList().size() + " persons";
-        // TODO: refine later
     }
 
     @Override
